+++ frontend/src/pages/scenes/fingerprintClusters/components/SceneChip.tsx
